renames infopopup to infotooltip
