+++ src/components/InfoTooltip/InfoTooltip.stories.tsx
@@ -1,18 +1,18 @@
 import React from 'react';
-import InfoPopup, { ArrowVariation } from './InfoPopup';
+import InfoTooltip, { ArrowVariation } from './InfoTooltip';
 
 export default {
-  title: 'Shared Components/InfoPopup',
-  component: InfoPopup,
+  title: 'Shared Components/InfoTooltip',
+  component: InfoTooltip,
 };
 
 const dummyText =
   'Explore our interactive U.S. COVID Map for the latest data on Cases, Vaccinations, Deaths, Positivity rate, and ICU capacity for your State, City, or County.';
 
 export const PointUpward = () => {
-  return <InfoPopup variation={ArrowVariation.DOWN} content={dummyText} />;
+  return <InfoTooltip variation={ArrowVariation.DOWN} content={dummyText} />;
 };
 
 export const PointDownward = () => {
-  return <InfoPopup variation={ArrowVariation.UP} content={dummyText} />;
+  return <InfoTooltip variation={ArrowVariation.UP} content={dummyText} />;
 };
